requirements.txt
streamlit>=1.28.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.17
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
sqlalchemy>=2.0.0
plotly>=5.15.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
